Trying to display nodes
